requirements.txt
asgiref==3.2.10
Django==3.1
django-cleanup==5.0.0
django-environ==0.4.5
django-filter==2.3.0
django-js-asset==1.2.2
django-mptt==0.11.0
djangorestframework==3.11.1
Markdown==3.2.2
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1

celery~=4.4.7
ip2geotools~=0.1.5